Conflicts:
	Makefile
	addons/packages/packetfence.spec
	docs/PacketFence_OpenDaylight_Install_Guide.asciidoc
	lib/pf/api.pm
	lib/pf/pfcmd.pm

+++ addons/packages/packetfence.spec
@@ -28,6 +28,7 @@
 #
 Summary: PacketFence network registration / worm mitigation system
 %global real_name packetfence
+%global perl_version 5.10.1
 Name: %{real_name}-source
 Version: %{ver}
 Release: %{rev}%{?dist}
@@ -96,7 +97,7 @@ Requires: make
 Requires: net-tools
 Requires: net-snmp >= 5.3.2.2
 Requires: mysql, mysql-server, perl(DBD::mysql)
-Requires: perl >= 5.8.8
+Requires: perl >= %{perl_version}
 # replaces the need for perl-suidperl which was deprecated in perl 5.12 (Fedora 14)
 Requires(pre): %{real_name}-pfcmd-suid
 Requires: perl(Bit::Vector)
@@ -128,6 +129,8 @@ Requires: perl(Time::HiRes)
 Requires: ipset, sudo
 Requires: perl(File::Which), perl(NetAddr::IP)
 Requires: perl(Net::LDAP)
+Requires: perl(Net::IP)
+Requires: perl(Digest::HMAC_MD5)
 # TODO: we should depend on perl modules not perl-libwww-perl package
 # find out what they are and specify them as perl(...::...) instead of perl-libwww-perl
 # LWP::Simple is one of them (required by inlined Net::MAC::Vendor and probably other stuff)
@@ -185,6 +188,8 @@ Requires: perl(UNIVERSAL::require)
 Requires: perl(YAML)
 Requires: perl(Try::Tiny)
 Requires: perl(Crypt::GeneratePassword)
+Requires: perl(Bytes::Random::Secure)
+Requires: perl(Crypt::Eksblowfish::Bcrypt)
 Requires: perl(MIME::Lite::TT)
 Requires: perl(Cache::Cache), perl(HTML::Parser)
 Requires: perl(URI::Escape::XS)
@@ -226,7 +231,7 @@ Requires: perl(File::Slurp)
 Requires: perl(Plack), perl(Plack::Middleware::ReverseProxy)
 Requires: perl(MooseX::Types::LoadableClass)
 Requires: perl(Moose) <= 2.1005
-Requires: perl(CHI) >= 0.56
+Requires: perl(CHI) >= 0.59
 Requires: perl(Data::Serializer)
 Requires: perl(Data::Structure::Util)
 Requires: perl(Data::Swap)
@@ -264,6 +269,10 @@ Requires: perl(Test::NoWarnings)
 Requires: perl(Net::UDP)
 # For managing the number of connections per device
 Requires: mod_qos
+Requires: %{real_name}-config = %{ver}
+Requires: %{real_name}-pfcmd-suid = %{ver}
+Requires: haproxy >= 1.5, keepalived >= 1.2
+Requires: fingerbank = 1.0.0
 
 %description -n %{real_name}
 
@@ -280,7 +289,7 @@ as
 
 %package -n %{real_name}-remote-snort-sensor
 Group: System Environment/Daemons
-Requires: perl >= 5.8.0, perl(File::Tail), perl(Config::IniFiles), perl(IO::Socket::SSL), perl(XML::Parser), perl(Crypt::SSLeay), perl(LWP::Protocol::https)
+Requires: perl >= %{perl_version}, perl(File::Tail), perl(Config::IniFiles), perl(IO::Socket::SSL), perl(XML::Parser), perl(Crypt::SSLeay), perl(LWP::Protocol::https), perl(SOAP::Lite)
 Requires: perl(Moo), perl(Data::MessagePack), perl(WWW::Curl)
 Conflicts: %{real_name}
 AutoReqProv: 0
@@ -295,7 +304,7 @@ server.
 
 %package -n %{real_name}-remote-arp-sensor
 Group: System Environment/Daemons
-Requires: perl >= 5.8.0, perl(Config::IniFiles), perl(IO::Socket::SSL), perl(XML::Parser), perl(Crypt::SSLeay), perl(LWP::Protocol::https), perl(Net::Pcap) >= 0.16, memcached, perl(Cache::Memcached)
+Requires: perl >= %{perl_version}, perl(Config::IniFiles), perl(IO::Socket::SSL), perl(XML::Parser), perl(Crypt::SSLeay), perl(LWP::Protocol::https), perl(Net::Pcap) >= 0.16, memcached, perl(Cache::Memcached)
 Requires: perl(Moo), perl(Data::MessagePack), perl(WWW::Curl)
 Conflicts: %{real_name}
 AutoReqProv: 0
@@ -310,13 +319,23 @@ for sending MAC and IP from ARP requests to a PacketFence server.
 %package -n %{real_name}-pfcmd-suid
 Group: System Environment/Daemons
 BuildRequires: gcc
-Requires: %{real_name} >= 3.6.0
 AutoReqProv: 0
 Summary: Replace pfcmd by a C wrapper for suid
 
 %description -n %{real_name}-pfcmd-suid
 The %{real_name}-pfcmd-suid is a C wrapper to replace perl-suidperl dependency.
 See https://bugzilla.redhat.com/show_bug.cgi?id=611009
+
+%package -n %{real_name}-config
+Group: System Environment/Daemons
+Requires: perl(Cache::BDB)
+Requires: perl(Log::Fast)
+AutoReqProv: 0
+Summary: Manage PacketFence Configuration
+BuildArch: noarch
+
+%description -n %{real_name}-config
+The %{real_name}-config is a daemon that manage PacketFence configuration.
 
 
 %prep
@@ -341,7 +360,7 @@ xsltproc -o docs/docbook/xsl/titlepage-fo.xsl \
     /usr/share/sgml/docbook/xsl-stylesheets/template/titlepage.xsl \
     docs/docbook/xsl/titlepage-fo.xml
 # admin, network device config, devel and ZEN install guides
-for GUIDE in PacketFence_Administration_Guide PacketFence_Developers_Guide PacketFence_Network_Devices_Configuration_Guide PacketFenceZEN_Installation_Guide PacketFence_Anyfi_Quick_Install_Guide PacketFence_Brocade_Quick_Install_Guide PacketFence_Cisco_Quick_Install_Guide PacketFence_Ruckus_Quick_Install_Guide PacketFence_OPSWAT_Quick_Install_Guide PacketFence_MobileIron_Quick_Install_Guide PacketFence_SEPM_Quick_Install_Guide PacketFence_PaloAlto_Quick_Install_Guide PacketFence_Barracuda_Quick_Install_Guide PacketFence_OpenDaylight_Install_Guide; do 
+for GUIDE in $(ls docs/PacketFence*.asciidoc | xargs -n1 -I'{}' basename '{}' .asciidoc) ;do
 asciidoc -a docinfo2 -b docbook -d book \
     -o docs/docbook/$GUIDE.docbook \
     docs/$GUIDE.asciidoc
@@ -381,9 +400,11 @@ done
 %{__install} -d $RPM_BUILD_ROOT/usr/local/pf/var/rrd 
 %{__install} -d $RPM_BUILD_ROOT/usr/local/pf/var/session
 %{__install} -d $RPM_BUILD_ROOT/usr/local/pf/var/webadmin_cache
+%{__install} -d $RPM_BUILD_ROOT/usr/local/pf/var/control
 touch $RPM_BUILD_ROOT/usr/local/pf/var/cache_control
 cp Makefile $RPM_BUILD_ROOT/usr/local/pf/
 cp -r bin $RPM_BUILD_ROOT/usr/local/pf/
+cp -r addons/pfconfig/ $RPM_BUILD_ROOT/usr/local/pf/addons/
 cp -r addons/captive-portal/ $RPM_BUILD_ROOT/usr/local/pf/addons/
 cp -r addons/dev-helpers/ $RPM_BUILD_ROOT/usr/local/pf/addons/
 cp -r addons/high-availability/ $RPM_BUILD_ROOT/usr/local/pf/addons/
@@ -434,6 +455,9 @@ cp -r README $RPM_BUILD_ROOT/usr/local/pf/
 cp -r README.network-devices $RPM_BUILD_ROOT/usr/local/pf/
 cp -r UPGRADE.asciidoc $RPM_BUILD_ROOT/usr/local/pf/
 cp -r UPGRADE.old $RPM_BUILD_ROOT/usr/local/pf/
+#pfconfig
+%{__install} -D -m0755 addons/pfconfig/pfconfig.init $RPM_BUILD_ROOT%{_initrddir}/packetfence-config
+#end pfconfig
 # logfiles
 for LOG in %logfiles; do
     touch $RPM_BUILD_ROOT%logdir/$LOG
@@ -471,6 +495,10 @@ ln -s ../sites-available/packetfence-soh packetfence-soh
 ln -s ../sites-available/packetfence-tunnel packetfence-tunnel
 ln -s ../sites-available/dynamic-clients dynamic-clients
 
+# Fingerbank symlinks
+cd $RPM_BUILD_ROOT/usr/local/pf/lib
+ln -s /usr/local/fingerbank/lib/fingerbank fingerbank
+
 cd $curdir
 #end create symlinks
 
@@ -483,6 +511,7 @@ if ! /usr/bin/id pf &>/dev/null; then
         /usr/sbin/useradd -r -d "/usr/local/pf" -s /bin/sh -c "PacketFence" -M pf || \
                 echo Unexpected error adding user "pf" && exit
 fi
+/usr/sbin/usermod -G fingerbank pf
 
 #if [ ! `tty | cut -c0-8` = "/dev/tty" ];
 #then
@@ -517,8 +546,15 @@ if ! /usr/bin/id pf &>/dev/null; then
                 echo Unexpected error adding user "pf" && exit
 fi
 
+%pre -n %{real_name}-config
+
+if ! /usr/bin/id pf &>/dev/null; then
+        /usr/sbin/useradd -r -d "/usr/local/pf" -s /bin/sh -c "PacketFence" -M pf || \
+                echo Unexpected error adding user "pf" && exit
+fi
+
+
 %post -n %{real_name}
-echo "Adding PacketFence startup script"
 /sbin/chkconfig --add packetfence
 
 #Check if log files exist and create them with the correct owner
@@ -536,6 +572,7 @@ if [ ! -f /usr/local/pf/conf/ssl/server.crt ]; then
     	-out /usr/local/pf/conf/ssl/server.crt\
     	-keyout /usr/local/pf/conf/ssl/server.key\
     	-nodes -config /usr/local/pf/conf/openssl.cnf
+    cat /usr/local/pf/conf/ssl/server.crt /usr/local/pf/conf/ssl/server.key > /usr/local/pf/conf/ssl/server.pem
 fi
 
 
@@ -567,6 +604,15 @@ if [ ! -f /usr/local/pf/raddb/certs/dh ]; then
   make dh
 else
   echo "DH already exists, won't touch it!"
+fi
+
+#Check if RADIUS have a dh
+if [ ! -f /usr/local/pf/conf/pf.conf ]; then
+  echo "Touch pf.conf because it doesnt exist"
+  touch /usr/local/pf/conf/pf.conf
+  chown pf.pf /usr/local/pf/conf/pf.conf
+else
+  echo "pf.conf already exists, won't touch it!"
 fi
 
 #Add for sudo 
@@ -608,6 +654,11 @@ echo "Adding PacketFence remote Snort Sensor startup script"
 echo "Adding PacketFence remote ARP Sensor startup script"
 /sbin/chkconfig --add pfarp
 
+%post -n %{real_name}-config
+chown pf.pf /usr/local/pf/conf/pfconfig.conf
+echo "Adding PacketFence config startup script"
+/sbin/chkconfig --add packetfence-config
+
 %preun -n %{real_name}
 if [ $1 -eq 0 ] ; then
         /sbin/service packetfence stop &>/dev/null || :
@@ -626,22 +677,41 @@ if [ $1 -eq 0 ] ; then
         /sbin/chkconfig --del pfarp
 fi
 
+%preun -n %{real_name}-config
+if [ $1 -eq 0 ] ; then
+        /sbin/service packetfence-config stop &>/dev/null || :
+        /sbin/chkconfig --del packetfence-config
+fi
+
 %postun -n %{real_name}
 if [ $1 -eq 0 ]; then
-        /usr/sbin/userdel pf || %logmsg "User \"pf\" could not be deleted."
-#       /usr/sbin/groupdel pf || %logmsg "Group \"pf\" could not be deleted."
-#else
-#       /sbin/service pf condrestart &>/dev/null || :
+        if /usr/bin/id pf &>/dev/null; then
+               /usr/sbin/userdel pf || %logmsg "User \"pf\" could not be deleted."
+#               /usr/sbin/groupdel pf || %logmsg "Group \"pf\" could not be deleted."
+#        else
+#               /sbin/service pf condrestart &>/dev/null || :
+        fi
 fi
 
 %postun -n %{real_name}-remote-snort-sensor
 if [ $1 -eq 0 ]; then
-        /usr/sbin/userdel pf || %logmsg "User \"pf\" could not be deleted."
+        if /usr/bin/id pf &>/dev/null; then
+                /usr/sbin/userdel pf || %logmsg "User \"pf\" could not be deleted."
+        fi
 fi
 
 %postun -n %{real_name}-remote-arp-sensor
 if [ $1 -eq 0 ]; then
-        /usr/sbin/userdel pf || %logmsg "User \"pf\" could not be deleted."
+        if /usr/bin/id pf &>/dev/null; then
+                /usr/sbin/userdel pf || %logmsg "User \"pf\" could not be deleted."
+        fi
+fi
+
+%postun -n %{real_name}-config
+if [ $1 -eq 0 ]; then
+        if /usr/bin/id pf &>/dev/null; then
+                /usr/sbin/userdel pf || %logmsg "User \"pf\" could not be deleted."
+        fi
 fi
 
 # TODO we should simplify this file manifest to the maximum keeping treating 
@@ -673,6 +743,10 @@ fi
                         /usr/local/pf/addons/logrotate
 %dir                    /usr/local/pf/addons/packages
                         /usr/local/pf/addons/packages/*
+%dir                    /usr/local/pf/addons/pfconfig
+%dir                    /usr/local/pf/addons/pfconfig/comparator
+%attr(0755, pf, pf)     /usr/local/pf/addons/pfconfig/comparator/*.pl
+%attr(0755, pf, pf)     /usr/local/pf/addons/pfconfig/comparator/*.sh
 %dir                    /usr/local/pf/addons/snort
 %attr(0755, pf, pf)     /usr/local/pf/addons/snort/update_rules.pl
                         /usr/local/pf/addons/snort/oinkmaster.conf
@@ -685,8 +759,11 @@ fi
 %attr(0755, pf, pf)     /usr/local/pf/addons/watchdog/*.sh
 %dir                    /usr/local/pf/bin
 %attr(0755, pf, pf)     /usr/local/pf/bin/pfcmd.pl
+%attr(0755, pf, pf)     /usr/local/pf/bin/pfcmd-old.pl
 %attr(0755, pf, pf)     /usr/local/pf/bin/pfcmd_vlan
 %attr(0755, pf, pf)     /usr/local/pf/bin/pftest
+%attr(0755, pf, pf)     /usr/local/pf/bin/cluster/management_update
+%attr(0755, pf, pf)     /usr/local/pf/bin/cluster/sync
 %doc                    /usr/local/pf/ChangeLog
 %dir                    /usr/local/pf/conf
                         /usr/local/pf/conf/*.example
@@ -756,12 +833,20 @@ fi
 %config(noreplace)      /usr/local/pf/conf/provisioning.conf
                         /usr/local/pf/conf/provisioning.conf.example
 %dir			/usr/local/pf/conf/radiusd
+%config(noreplace)      /usr/local/pf/conf/radiusd/clients.conf.inc
+                        /usr/local/pf/conf/radiusd/clients.conf.inc.example
+%config(noreplace)      /usr/local/pf/conf/radiusd/packetfence-cluster
+                        /usr/local/pf/conf/radiusd/packetfence-cluster.example
+%config(noreplace)      /usr/local/pf/conf/radiusd/proxy.conf.inc
+                        /usr/local/pf/conf/radiusd/proxy.conf.inc.example
 %config(noreplace)	/usr/local/pf/conf/radiusd/eap.conf
                         /usr/local/pf/conf/radiusd/eap.conf.example
 %config(noreplace)	/usr/local/pf/conf/radiusd/radiusd.conf
                         /usr/local/pf/conf/radiusd/radiusd.conf.example
 %config(noreplace)	/usr/local/pf/conf/radiusd/sql.conf
                         /usr/local/pf/conf/radiusd/sql.conf.example
+%config(noreplace)      /usr/local/pf/conf/realm.conf
+                        /usr/local/pf/conf/realm.conf.example
 %dir                    /usr/local/pf/conf/snort
 %config(noreplace)      /usr/local/pf/conf/snort/classification.config
                         /usr/local/pf/conf/snort/classification.config.example
@@ -775,6 +860,8 @@ fi
 %config(noreplace)      /usr/local/pf/conf/vlan_filters.conf
                         /usr/local/pf/conf/vlan_filters.conf.example
 %config                 /usr/local/pf/conf/dhcpd.conf
+%config(noreplace)      /usr/local/pf/conf/haproxy.conf
+                        /usr/local/pf/conf/haproxy.conf.example
 %dir                    /usr/local/pf/conf/httpd.conf.d
 %config                 /usr/local/pf/conf/httpd.conf.d/captive-portal-common.conf
 %config                 /usr/local/pf/conf/httpd.conf.d/httpd.aaa
@@ -787,6 +874,10 @@ fi
 %config(noreplace)	/usr/local/pf/conf/httpd.conf.d/ssl-certificates.conf
                         /usr/local/pf/conf/httpd.conf.d/ssl-certificates.conf.example
 %config(noreplace)      /usr/local/pf/conf/iptables.conf
+%config(noreplace)      /usr/local/pf/conf/keepalived.conf
+                        /usr/local/pf/conf/keepalived.conf.example
+%config(noreplace)      /usr/local/pf/conf/cluster.conf
+                        /usr/local/pf/conf/cluster.conf.example
 %config(noreplace)      /usr/local/pf/conf/listener.msg
                         /usr/local/pf/conf/listener.msg.example
 %config(noreplace)      /usr/local/pf/conf/popup.msg
@@ -831,6 +922,7 @@ fi
                         /usr/local/pf/html/captive-portal/content/guest-management.js
                         /usr/local/pf/html/captive-portal/content/timerbar.js
                         /usr/local/pf/html/captive-portal/content/shared_mdm_profile.mobileconfig
+                        /usr/local/pf/html/captive-portal/content/packetfence-windows-agent.exe
 %dir                    /usr/local/pf/html/captive-portal/content/images
                         /usr/local/pf/html/captive-portal/content/images/*
 %dir                    /usr/local/pf/html/captive-portal/lib
@@ -869,7 +961,40 @@ fi
 %dir                    /usr/local/pf/html/common
                         /usr/local/pf/html/common/*
                         /usr/local/pf/html/pfappserver/
+%config(noreplace)      /usr/local/pf/html/pfappserver/pfappserver.conf
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Admin.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/AdminRoles.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/Authentication.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/Authentication/Source.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/Firewall_SSO.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/FloatingDevice.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/MacAddress.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/Networks.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/Pf.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/Profile/Default.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/Profile.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/Provisioning.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/Realm.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/Switch.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/System.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Configuration.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Configurator.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/UserAgents.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/Wrix.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/DB.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Graph.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Interface.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Node.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Roles.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Root.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/SavedSearch/Node.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/SavedSearch/User.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Service.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/SoH.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/User.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Violation.pm
                         /usr/local/pf/lib
+%exclude                /usr/local/pf/lib/pfconfig*
 %config(noreplace)      /usr/local/pf/lib/pf/billing/custom.pm
 %config(noreplace)      /usr/local/pf/lib/pf/floatingdevice/custom.pm
 %config(noreplace)      /usr/local/pf/lib/pf/inline/custom.pm
@@ -948,6 +1073,7 @@ fi
 %dir                    /usr/local/pf/var/rrd
 %dir                    /usr/local/pf/var/session
 %dir                    /usr/local/pf/var/webadmin_cache
+%dir                    /usr/local/pf/var/control
 %config(noreplace)      /usr/local/pf/var/cache_control
 
 # Remote snort sensor file list
@@ -977,7 +1103,39 @@ fi
 %files -n %{real_name}-pfcmd-suid
 %attr(6755, root, root) /usr/local/pf/bin/pfcmd
 
+%files -n %{real_name}-config
+%attr(0755, root, root) %{_initrddir}/packetfence-config
+%dir                    /usr/local/pf
+%dir                    /usr/local/pf/conf
+%config(noreplace)      /usr/local/pf/conf/pfconfig.conf
+%dir                    /usr/local/pf/lib
+%dir                    /usr/local/pf/lib/pfconfig
+                        /usr/local/pf/lib/pfconfig/*
+%attr(0755, pf, pf)     /usr/local/pf/sbin/pfconfig
+%dir                    /usr/local/pf/addons/pfconfig
+%attr(0755, pf, pf)     /usr/local/pf/addons/pfconfig/cmd.pl
+%exclude                /usr/local/pf/addons/pfconfig/README.asciidoc
+%exclude                /usr/local/pf/addons/pfconfig/pfconfig.init
+
 %changelog
+* Wed Apr 22 2015 Inverse <info@inverse.ca> - 5.0.1-1
+- New release 5.0.1
+
+* Wed Apr 15 2015 Inverse <info@inverse.ca> - 5.0.0-1
+- New release 5.0.0
+
+* Fri Mar 06 2015 Inverse <info@inverse.ca> - 4.7.0-1
+- New release 4.7.0
+
+* Thu Feb 19 2015 Inverse <info@inverse.ca> - 4.6.1-1
+- New release 4.6.1
+
+* Wed Feb 04 2015 Inverse <info@inverse.ca> - 4.6.0-1
+- New release 4.6.0
+
+* Mon Nov 10 2014 Inverse <info@inverse.ca> - 4.5.1-1
+- New release 4.5.1
+
 * Wed Oct 22 2014 Inverse <info@inverse.ca> - 4.5.0-1
 - New release 4.5.0
